Refactor and modularize
